Data logging added. Tables and figures annot solved.
Added creation of new folder with info on all params in which the csv and png files
are saved.
Currently saving the tissue window iamges, h5ad file of tissue after clustering and
images of clustering and comaprisons of cell mixtures and its clusters.
Table of cell mixture parcenatages is also saved